Add files via upload
The problem situation is inspired by the "Eurobot" contest. The objective of the robot is to grab and bring back targets to its team basis. In collaboration with 4 other students, different modules such as localization, path-planning and low-level control have been implemented. The mobile robot controller was programmed in C. The employe simulation environment is Robotran.